Donut/requirements.txt
transformers==4.49.0
torch==2.6.0
torchvision==0.21.0
sentencepiece==0.2.0
Pillow==11.1.0